公会功能 离线玩家处理接口

+++ main.go
@@ -61,21 +61,23 @@ func main() {
 	// 监听系统的kill信号
 	signalKillNotify := make(chan os.Signal, 1)
 	signal.Notify(signalKillNotify, os.Interrupt, os.Kill, syscall.SIGTERM)
-	// 加一个控制台输入,以方便调试
-	go func() {
-		consoleReader := bufio.NewReader(os.Stdin)
-		for {
-			lineBytes, _, _ := consoleReader.ReadLine()
-			line := strings.ToLower(string(lineBytes))
-			logger.Info("line:%v", line)
-			if line == "close" || line == "exit" {
-				logger.Info("kill by console input")
-				// 模拟一个kill信号,以方便测试服务器退出流程
-				signalKillNotify <- os.Kill
-				return
+	if !isDaemon {
+		// 加一个控制台输入,以方便调试
+		go func() {
+			consoleReader := bufio.NewReader(os.Stdin)
+			for {
+				lineBytes, _, _ := consoleReader.ReadLine()
+				line := strings.ToLower(string(lineBytes))
+				logger.Info("line:%v", line)
+				if line == "close" || line == "exit" {
+					logger.Info("kill by console input")
+					// 模拟一个kill信号,以方便测试服务器退出流程
+					signalKillNotify <- os.Kill
+					return
+				}
 			}
-		}
-	}()
+		}()
+	}
 	// 阻塞等待系统关闭信号
 	logger.Info("wait for kill signal")
 	select {
